+++ net/minecraft/entity/DataWatcher.java
@@ -46,13 +46,13 @@ public class DataWatcher
 		{
 			throw new IllegalArgumentException("Data value id is too big with " + id + "! (Max is " + 31 + ")");
 		}
-		if (this.watchedObjects.containsKey(Integer.valueOf(id)))
+		if (this.watchedObjects.containsKey(id))
 		{
 			throw new IllegalArgumentException("Duplicate id value for " + id + "!");
 		}
 		DataWatcher.WatchableObject datawatcher$watchableobject = new DataWatcher.WatchableObject(integer.intValue(), id, object);
 		this.lock.writeLock().lock();
-		this.watchedObjects.put(Integer.valueOf(id), datawatcher$watchableobject);
+		this.watchedObjects.put(id, datawatcher$watchableobject);
 		this.lock.writeLock().unlock();
 		this.isBlank = false;
 	}
@@ -64,7 +64,7 @@ public class DataWatcher
     {
         DataWatcher.WatchableObject datawatcher$watchableobject = new DataWatcher.WatchableObject(type, id, (Object)null);
         this.lock.writeLock().lock();
-        this.watchedObjects.put(Integer.valueOf(id), datawatcher$watchableobject);
+        this.watchedObjects.put(id, datawatcher$watchableobject);
         this.lock.writeLock().unlock();
         this.isBlank = false;
     }
@@ -121,13 +121,13 @@ public class DataWatcher
 
         try
         {
-            datawatcher$watchableobject = (DataWatcher.WatchableObject)this.watchedObjects.get(Integer.valueOf(id));
+            datawatcher$watchableobject = (DataWatcher.WatchableObject)this.watchedObjects.get(id);
         }
         catch (Throwable throwable)
         {
             CrashReport crashreport = CrashReport.makeCrashReport(throwable, "Getting synched entity data");
             CrashReportCategory crashreportcategory = crashreport.makeCategory("Synched entity data");
-            crashreportcategory.addCrashSection("Data ID", Integer.valueOf(id));
+            crashreportcategory.addCrashSection("Data ID", id);
             throw new ReportedException(crashreport);
         }
 
@@ -315,19 +315,19 @@ public class DataWatcher
             switch (j)
             {
                 case 0:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Byte.valueOf(buffer.readByte()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readByte());
                     break;
 
                 case 1:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Short.valueOf(buffer.readShort()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readShort());
                     break;
 
                 case 2:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Integer.valueOf(buffer.readInt()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readInt());
                     break;
 
                 case 3:
-                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, Float.valueOf(buffer.readFloat()));
+                    datawatcher$watchableobject = new DataWatcher.WatchableObject(j, k, buffer.readFloat());
                     break;
 
                 case 4:
@@ -364,7 +364,7 @@ public class DataWatcher
 
         for (DataWatcher.WatchableObject datawatcher$watchableobject : p_75687_1_)
         {
-            DataWatcher.WatchableObject datawatcher$watchableobject1 = (DataWatcher.WatchableObject)this.watchedObjects.get(Integer.valueOf(datawatcher$watchableobject.getDataValueId()));
+            DataWatcher.WatchableObject datawatcher$watchableobject1 = (DataWatcher.WatchableObject)this.watchedObjects.get(datawatcher$watchableobject.getDataValueId());
 
             if (datawatcher$watchableobject1 != null)
             {
@@ -389,14 +389,14 @@ public class DataWatcher
 
     static
     {
-        dataTypes.put(Byte.class, Integer.valueOf(0));
-        dataTypes.put(Short.class, Integer.valueOf(1));
-        dataTypes.put(Integer.class, Integer.valueOf(2));
-        dataTypes.put(Float.class, Integer.valueOf(3));
-        dataTypes.put(String.class, Integer.valueOf(4));
-        dataTypes.put(ItemStack.class, Integer.valueOf(5));
-        dataTypes.put(BlockPos.class, Integer.valueOf(6));
-        dataTypes.put(Rotations.class, Integer.valueOf(7));
+        dataTypes.put(Byte.class, 0);
+        dataTypes.put(Short.class, 1);
+        dataTypes.put(Integer.class, 2);
+        dataTypes.put(Float.class, 3);
+        dataTypes.put(String.class, 4);
+        dataTypes.put(ItemStack.class, 5);
+        dataTypes.put(BlockPos.class, 6);
+        dataTypes.put(Rotations.class, 7);
     }
 
     public static class WatchableObject
