+++ net/minecraft/entity/item/EntityArmorStand.java
@@ -79,7 +79,7 @@ public class EntityArmorStand extends EntityLivingBase
     protected void entityInit()
     {
         super.entityInit();
-        this.dataWatcher.addObject(10, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(10, (byte) 0);
         this.dataWatcher.addObject(11, DEFAULT_HEAD_ROTATION);
         this.dataWatcher.addObject(12, DEFAULT_BODY_ROTATION);
         this.dataWatcher.addObject(13, DEFAULT_LEFTARM_ROTATION);
@@ -800,7 +800,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -2);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean isSmall()
@@ -821,7 +821,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -3);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean hasNoGravity()
@@ -842,7 +842,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -5);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean getShowArms()
@@ -863,7 +863,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -9);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean hasNoBasePlate()
@@ -884,7 +884,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -17);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean func_181026_s()
