+++ net/minecraft/entity/monster/EntityGhast.java
@@ -46,7 +46,7 @@ public class EntityGhast extends EntityFlying implements IMob
 
     public void setAttacking(boolean p_175454_1_)
     {
-        this.dataWatcher.updateObject(16, Byte.valueOf((byte)(p_175454_1_ ? 1 : 0)));
+        this.dataWatcher.updateObject(16, (byte) (p_175454_1_ ? 1 : 0));
     }
 
     public int getFireballStrength()
@@ -88,7 +88,7 @@ public class EntityGhast extends EntityFlying implements IMob
     protected void entityInit()
     {
         super.entityInit();
-        this.dataWatcher.addObject(16, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(16, (byte) 0);
     }
 
     protected void applyEntityAttributes()
