+++ net/minecraft/entity/projectile/EntityArrow.java
@@ -507,9 +507,9 @@ public class EntityArrow extends Entity implements IProjectile {
 		byte b0 = this.dataWatcher.getWatchableObjectByte(16);
 
 		if (critical) {
-			this.dataWatcher.updateObject(16, Byte.valueOf((byte) (b0 | 1)));
+			this.dataWatcher.updateObject(16, (byte) (b0 | 1));
 		} else {
-			this.dataWatcher.updateObject(16, Byte.valueOf((byte) (b0 & -2)));
+			this.dataWatcher.updateObject(16, (byte) (b0 & -2));
 		}
 	}
