+++ net/minecraft/world/storage/MapStorage.java
@@ -168,7 +168,7 @@ public class MapStorage
                     {
                         NBTTagShort nbttagshort = (NBTTagShort)nbtbase;
                         short short1 = nbttagshort.getShort();
-                        this.idCounts.put(s, Short.valueOf(short1));
+                        this.idCounts.put(s, short1);
                     }
                 }
             }
@@ -188,11 +188,11 @@ public class MapStorage
 
         if (oshort == null)
         {
-            oshort = Short.valueOf((short)0);
+            oshort = (short) 0;
         }
         else
         {
-            oshort = Short.valueOf((short)(oshort.shortValue() + 1));
+            oshort = (short) (oshort.shortValue() + 1);
         }
 
         this.idCounts.put(key, oshort);
